modified kafka config to add keys and better event handling

+++ BeerQuery/src/main/java/beer/api/beer/query/listeners/BeerDeletedEventListener.java
@@ -4,7 +4,10 @@ import beer.api.beer.query.events.BeerDeletedEvent;
 import beer.api.beer.query.repositories.BeerRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.kafka.support.KafkaHeaders;
+import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -18,10 +21,19 @@ public class BeerDeletedEventListener {
         this.beerRepository = beerRepository;
     }
 
-    @KafkaListener(topics = "beer-topic", groupId = "beer-consumers", containerFactory = "kafkaListenerContainerFactory")
-    public void handleBeerDeletedEvent(BeerDeletedEvent event) {
+    @Value("${spring.kafka.event.key.beer.deleted}")
+    private String BEER_DELETED_KEY;
+
+    @KafkaListener(topics = "beer-topic", groupId = "beer-consumers", containerFactory = "beerDeletedListenerFactory")
+    public void handleBeerDeletedEvent(BeerDeletedEvent event, @Header(KafkaHeaders.RECEIVED_KEY) String key) {
 
         try {
+
+            if (!BEER_DELETED_KEY.equals(key)) {
+                log.warn("Received event with unexpected key: {}", key);
+                return;
+            }
+
             log.info("Received event {}", event);
 
             beerRepository.deleteById(String.valueOf(event.getBeerId()));
